♻️ (aa-auth): Move ProblemDetail into aa-auth
Move the self-contained RFC-7807 `ProblemDetail` type (and its tests) from
`aa-api/src/error.rs` into the `aa-auth` leaf crate via `git mv`. The auth
framework's `AuthError::into_response` renders `ProblemDetail`, so it must live
beside the auth code in the leaf. `aa-api/src/error.rs` becomes a one-line
re-export (`pub use aa_auth::ProblemDetail;`) so all ~27 `crate::error::
ProblemDetail` sites and the `openapi.rs` schema registration keep resolving
unchanged. The derived `ToSchema` is identical, so the OpenAPI spec is
byte-identical.

Refs AAASM-3905
